Add files via upload
This is just a file with the code to read the different data sets we have on Teams. At the end of the document, there are some pointers on where to start. I mainly addressed the question about crime vs occupation. Please feel free to add any comments about house prices vs crime.